UAF-259 1) Redefine UAIdentityManagementPersonDocument entity inheritance with EclipseLink JPA. OJB is no longer used for rice-impl by rice. 2) Modified IdentityManagementPersonDocument with removing attributes tag which were defined for KFS searching attributes not intended for RICE use
UAF-259 refactoring edu new class comments

UAF-259 refactoring per code review commit
