+++ rice-middleware/impl/src/main/resources/org/kuali/rice/kim/document/datadictionary/IdentityManagementPersonDocument.xml
@@ -21,7 +21,7 @@
   <bean id="IdentityManagementPersonDocument-parentBean" abstract="true" parent="TransactionalDocumentEntry">
     <property name="documentTypeName" value="IdentityManagementPersonDocument"/>
     <!-- UA changes -->
-    <property name="documentClass" value="edu.arizona.kim.document.UaIdentityManagementPersonDocument"/>
+    <property name="documentClass" value="edu.arizona.rice.kim.document.UaIdentityManagementPersonDocument"/>
     <!-- end -->
     <property name="businessRulesClass" value="org.kuali.rice.kim.document.rule.IdentityManagementPersonDocumentRule"/>
     <property name="documentAuthorizerClass" value="org.kuali.rice.kim.document.authorization.IdentityManagementKimDocumentAuthorizer"/>
