+++ rice-middleware/kim/kim-ldap/src/main/java/edu/arizona/kim/ojb-kim.xml
@@ -8,8 +8,9 @@
 		<object-cache class="org.apache.ojb.broker.cache.ObjectCachePerBrokerImpl" />
 	</jdbc-connection-descriptor>
 
+  <!-- UA KFS7 upgrade, retired OJB mapping and replaced by JPA eclipselink -->
   <!-- Ua PersonDocument -->
-  <class-descriptor class="edu.arizona.kim.document.UaIdentityManagementPersonDocument" table="KRIM_PERSON_DOCUMENT_T">
+  <class-descriptor class="edu.arizona.rice.kim.document.UaIdentityManagementPersonDocument" table="KRIM_PERSON_DOCUMENT_T">
 		<field-descriptor name="documentNumber" column="FDOC_NBR"     jdbc-type="VARCHAR" primarykey="true" />
 		<field-descriptor name="entityId"       column="entity_id"    jdbc-type="VARCHAR" autoincrement="true" sequence-name="KRIM_ENTITY_ID_S" />
 		<field-descriptor name="principalId"    column="prncpl_id"    jdbc-type="VARCHAR" autoincrement="true" sequence-name="KRIM_PRNCPL_ID_S" />
